feat(webui): code-split remaining team and share sub-routes to optimize vendor chunk size
